app/requirements.txt
click==8.1.3
dnspython==2.3.0
Flask==2.2.2
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
Werkzeug==2.2.2
WTForms==3.0.1
